update Dockerfile,rickmenu have some problems
