+++ app.py
@@ -242,4 +242,4 @@ def handle_message(event):
 
 
 if __name__ == "__main__":
-    app.run(debug=True, host="0.0.0.0", port=8787)
+    app.run(debug=True, host="0.0.0.0", port=8080)
