+++ feature/line/__init__.py
@@ -1,3 +1,3 @@
-from .rich_manu import RichMenuManager
+from .rich_menu import RichMenuManager
 
 __all__ = ["RichMenuManager",]
